requirements.txt
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
diffusers==0.30.3
transformers==4.44.2
huggingface-hub==0.25.2
Flask==3.1.2
soundfile==0.13.1
numpy<2.0
scipy>=1.0.0
opencv-python==4.11.0.86
pillow==10.4.0
requests==2.32.5
accelerate==1.11.0
xformers==0.0.23
ultralytics==8.2.32
piper-tts==1.3.0
onnxruntime-gpu==1.23.2
googletrans==4.0.0rc1
